The 'process.py' module is now turned into several smaller ones, each containing only 1 function.
